[Task#002]-[User Configuration and logs ]
