--- src/main/java/com/example/springbootandmongodb/UserConfig.java
@@ -1,7 +1,0 @@
-package com.example.springbootandmongodb;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class UserConfig {
-}
